@@ -1,40 +1,62 @@
-import { Module } from '@nuxt/types';
-const path = require('path');
+import { addImportsDir, createResolver, defineNuxtModule } from '@nuxt/kit';
+import { defu } from 'defu';
 
-export interface Config {
-  apiKey: string;
+type TargetParam = 'server' | 'all';
+
+export interface ModuleOptions {
   serviceDomain: string;
+  apiKey: string;
+  target?: TargetParam;
 }
 
-export interface Options {
-  fileName?: string;
-  mode?: 'server' | 'client' | 'all';
-  options: Config;
+export default defineNuxtModule<ModuleOptions>({
+  meta: {
+    name: 'nuxt-microcms-module',
+    configKey: 'microCMS',
+    compatibility: {
+      nuxt: '^3.0.0',
+    },
+  },
+  setup(options, nuxt) {
+    // Validate options
+    if (!options.serviceDomain) {
+      throw new Error('serviceDomain is required');
+    }
+    if (!options.apiKey) {
+      throw new Error('apiKey is required');
+    }
+
+    options.target = options.target || 'server';
+
+    // Add runtime config
+    nuxt.options.runtimeConfig.microCMS = defu(
+      nuxt.options.runtimeConfig.microCMS,
+      options
+    );
+    nuxt.options.runtimeConfig.public = nuxt.options.runtimeConfig.public || {};
+    nuxt.options.runtimeConfig.public.microCMS =
+      nuxt.options.runtimeConfig.public.microCMS || {};
+    nuxt.options.runtimeConfig.public.microCMS.serviceDomain =
+      nuxt.options.runtimeConfig.public.microCMS.serviceDomain ||
+      nuxt.options.runtimeConfig.microCMS.serviceDomain;
+
+    if (nuxt.options.dev || options.target === 'all') {
+      nuxt.options.runtimeConfig.public.microCMS = defu(
+        nuxt.options.runtimeConfig.public.microCMS,
+        nuxt.options.runtimeConfig.microCMS
+      );
+    }
+
+    const { resolve } = createResolver(import.meta.url);
+
+    // Add composables
+    addImportsDir(resolve('./runtime/composables'));
+  },
+});
+
+declare module '@nuxt/schema' {
+  interface RuntimeConfig {
+    serviceDomain: string;
+    apiKey: string;
+  }
 }
-
-interface Package {
-  name: string;
-  version: string;
-  description: string;
-  main: string;
-  license: string;
-}
-
-export interface ModuleWithMeta<T> extends Module<T> {
-  meta: Package;
-}
-
-const microcmsModule: ModuleWithMeta<Options> = function (this, moduleOptions) {
-  const _options = { ...this.options.microcms, ...moduleOptions };
-  const { fileName, mode, options } = _options;
-  this.addPlugin({
-    src: path.resolve(__dirname, 'plugin.js'),
-    fileName: fileName || 'microcms.js',
-    mode,
-    options
-  });
-};
-
-microcmsModule.meta = require('../package.json');
-
-export default microcmsModule;
